Add Reader.WithBufferSize() and rework unit tests
This adds a Reader.WithBufferSize() configuration function that calls
`Buffer()` on the `bufio.Scanner()` in the parser, allowing parsing of
large TSV records. The default is a 64K buffer, but zeek has no
limitation of TSV colunm value sizes, so we could easily hit
`bufio.Scanner: token too long`.

The unit tests have been updated with a case that fails with the default
buffer size, and a case that passes with the same input but a larger
(configured buffer).

While I was working on the unit tests I realised that my previous
implementation of `collectWithErrors` didn't make sense: once we hit an
error it should return, not "accumulate errors". Now we have
`collectWithError` that returns only a slice of `Records` + `error`,
which is both correct and it makes `MakeReadTester()` quite a bit
simpler too.