Make Reader support seeking without breaking v1 API
This implements what I intended for v1.1.0, but without breaking the v1
API (fixed in v1.1.1 which reverts the breaking change):

Like `NewLazySeekableReader()`, the `NewSeekableReader()` function added
here returns a `Reader` that supports the (previously added) `Seek()`
method. If the `Reader` is obtained from `NewReader()` instead, calling
`Seek()` will return `ErrSeekingUnsupported`.

Most of the logic for this is already in the `Parser`, so this just
provides a non-API breaking way to get the new functionality (whcih
requires a `io.ReadSeeker` input stream (vs just a `io.Reader`).